+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add `MyBulkSign.bulkSignatures` endpoint
+- Add `MyBulkSign.envelopes` endpoint
+- Add `UserBulkSignEnvelope` resource
 - Fix `MyBulkSignatures.sign` request method
 - Add `MyBulkSignatures.sign` resource
 - Add `MyEnvelope.bulkSign` resource